Solved Silver II 2022-09-14

+++ 1654/1654.cpp
@@ -17,12 +17,12 @@ int main() {
         if(maxi < arr[i]) maxi = arr[i];
     }
 
-    int left = 1;
-    int right = maxi;
-    int ans = 0;
+    long long left = 1;
+    long long right = maxi;
+    long long ans = 0;
     while(left <= right) {
-        int mid = (left + right) / 2;
-        int lan = 0;
+        long long mid = (left + right) / 2;
+        long long lan = 0;
         for(int i = 0; i < k; i++) lan += (arr[i] / mid);
         if(lan < n) {
             right = mid - 1;
